refactor(link):optimize kafka send
