+++ core/plugin/link/service/community/tools/http/execution_server.py
@@ -8,15 +8,11 @@ It handles authentication, parameter validation, and response processing.
 import base64
 import json
 import os
-import queue
-import threading
-import time
 from typing import Any, Dict, List, Optional, Tuple
 
 from common.otlp.log_trace.node_trace_log import NodeTraceLog, Status
 from common.otlp.metrics.meter import Meter
 from common.otlp.trace.span import Span
-from common.service import get_kafka_producer_service
 from loguru import logger
 from opentelemetry.trace import Status as OTelStatus
 from opentelemetry.trace import StatusCode
@@ -31,6 +27,7 @@ from plugin.link.api.schemas.community.tools.http.execution_schema import (
 from plugin.link.consts import const
 from plugin.link.domain.models.manager import get_db_engine
 from plugin.link.exceptions.sparklink_exceptions import SparkLinkBaseException
+from plugin.link.infra.kafka_telemetry import send_telemetry_sync
 from plugin.link.infra.tool_crud.process import ToolCrudOperation
 from plugin.link.infra.tool_exector.process import HttpRun
 from plugin.link.utils.errors.code import ErrCode
@@ -50,73 +47,6 @@ default_value = {
     " 'boolean'": False,
     " 'integer'": 0,
 }
-
-global_kafka_queue: queue.Queue = queue.Queue(maxsize=10000)
-
-KAFKA_MAX_WORKERS = 10
-KAFKA_WORKER_TIMEOUT = 30
-KAFKA_WATCHDOG_INTERVAL = 5
-
-_worker_threads = []
-_worker_last_active = []
-_worker_lock = threading.Lock()
-
-
-def init_kafka_send_workers() -> None:
-    """Initialize Kafka producer workers"""
-    for idx in range(KAFKA_MAX_WORKERS):
-        thread = threading.Thread(target=_kafka_worker_func, args=(idx,), daemon=True)
-        with _worker_lock:
-            _worker_threads.append(thread)
-            _worker_last_active.append(time.time())
-
-    watchdog_thread = threading.Thread(target=_kafka_watchdog_func, daemon=True)
-    watchdog_thread.start()
-
-
-def _kafka_worker_func(worker_idx: int) -> None:
-    """Kafka worker function"""
-
-    kafka_producer = None
-    while True:
-        try:
-            if not kafka_producer:
-                kafka_producer = get_kafka_producer_service()
-
-            with _worker_lock:
-                _worker_last_active[worker_idx] = time.time()
-            logger.info(f"[Worker {worker_idx}] Waiting for data to send to kafka")
-            data = global_kafka_queue.get(timeout=1)
-            kafka_producer.send(os.getenv(const.KAFKA_TOPIC_KEY), data)
-
-        except queue.Empty:
-            time.sleep(1)
-        except Exception as e:
-            logger.error(f"[Worker {worker_idx}] Failed to send data to kafka: {e}")
-            kafka_producer = None
-        finally:
-            continue
-
-
-def _kafka_watchdog_func() -> None:
-    """Watchdog to monitor worker threads and restart if stuck"""
-
-    while True:
-        time.sleep(KAFKA_WATCHDOG_INTERVAL)
-        now = time.time()
-
-        with _worker_lock:
-            for idx, last_active in enumerate(_worker_last_active):
-                if now - last_active > KAFKA_WORKER_TIMEOUT:
-                    logger.error(f"[Watchdog] Worker {idx} seems stuck, restarting")
-
-                    thread = threading.Thread(
-                        target=_kafka_worker_func, args=(idx,), daemon=True
-                    )
-                    logger.info(f"[Watchdog] Starting worker {idx}")
-                    thread.start()
-                    _worker_threads[idx] = thread
-                    _worker_last_active[idx] = now
 
 
 def extract_request_params(
@@ -141,17 +71,6 @@ def extract_request_params(
     return app_id, uid, caller
 
 
-async def send_telemetry(node_trace: NodeTraceLog) -> None:
-    """Send telemetry data to Kafka."""
-    if os.getenv(const.OTLP_ENABLE_KEY, "0").lower() == "1":
-        node_trace.start_time = int(round(time.time() * 1000))
-        try:
-            logger.debug(f"Current kafka queue size: {global_kafka_queue.qsize()}")
-            global_kafka_queue.put(node_trace.to_json(), block=False)
-        except queue.Full:
-            logger.debug("Kafka queue is full, drop telemetry data")
-
-
 async def handle_validation_error(
     validate_err: str, span_context: Span, node_trace: NodeTraceLog, m: Meter
 ) -> HttpRunResponse:
@@ -163,7 +82,7 @@ async def handle_validation_error(
             code=ErrCode.JSON_PROTOCOL_PARSER_ERR.code,
             message=validate_err,
         )
-        await send_telemetry(node_trace)
+        send_telemetry_sync(node_trace)
 
     return HttpRunResponse(
         header=HttpRunResponseHeader(
@@ -197,7 +116,7 @@ async def handle_sparklink_error(
             code=err.code,
             message=err.message,
         )
-        await send_telemetry(node_trace)
+        send_telemetry_sync(node_trace)
 
     return HttpRunResponse(
         header=HttpRunResponseHeader(
@@ -230,7 +149,7 @@ async def handle_custom_error(
             code=error_code.code,
             message=message,
         )
-        await send_telemetry(node_trace)
+        send_telemetry_sync(node_trace)
 
     return HttpRunResponse(
         header=HttpRunResponseHeader(
@@ -264,7 +183,7 @@ async def handle_general_exception(
             code=ErrCode.COMMON_ERR.code,
             message=f"{ErrCode.COMMON_ERR.msg}: {err}",
         )
-        await send_telemetry(node_trace)
+        send_telemetry_sync(node_trace)
 
     return HttpRunResponse(
         header=HttpRunResponseHeader(
@@ -373,7 +292,7 @@ async def handle_success_response(
             code=ErrCode.SUCCESSES.code,
             message=ErrCode.SUCCESSES.msg,
         )
-        await send_telemetry(node_trace)
+        send_telemetry_sync(node_trace)
 
     return HttpRunResponse(
         header=HttpRunResponseHeader(
@@ -412,7 +331,7 @@ async def handle_debug_validation_error(
             code=ErrCode.JSON_PROTOCOL_PARSER_ERR.code,
             message=validate_err,
         )
-        await send_telemetry(node_trace)
+        send_telemetry_sync(node_trace)
 
     return HttpRunResponse(
         header=HttpRunResponseHeader(
@@ -442,7 +361,7 @@ async def handle_debug_success_response(
             code=ErrCode.SUCCESSES.code,
             message=ErrCode.SUCCESSES.msg,
         )
-        await send_telemetry(node_trace)
+        send_telemetry_sync(node_trace)
 
     return ToolDebugResponse(
         header=ToolDebugResponseHeader(
